API

Modules

DoubleBetaDecayGenerators.Legacy

Types and constants

DoubleBetaDecayGenerators.TwoNuDBDData

DoubleBetaDecayGenerators.ZeroNuDBDData

Functions and macros

DoubleBetaDecayGenerators.Legacy.dk0gen

DoubleBetaDecayGenerators.Legacy.momentum

DoubleBetaDecayGenerators.Legacy.rand_2nbb

DoubleBetaDecayGenerators.acute_inv_cdf

Documentation

DoubleBetaDecayGenerators.TwoNuDBDData — Type

TwoNuDBDData(path)

A structure to hold numerical calculations of two-neutrino double-beta decay in the format provided by Iachello, F. and Kotila, J. [1].

[1] https://nucleartheory.yale.edu/double-beta-decay-phase-space-factors

DoubleBetaDecayGenerators.ZeroNuDBDData — Type

ZeroNuDBDData(path)

A structure to hold numerical calculations of zero-neutrino double-beta decay in the format provided by Iachello, F. and Kotila, J. [1].

[1] https://nucleartheory.yale.edu/double-beta-decay-phase-space-factors

DoubleBetaDecayGenerators.acute_inv_cdf — Method

acute_inv_cdf(x, q)

The inverse of the cumulative of the "acute" probability distribution, where q is the linear angular coefficient.

[1] https://stats.stackexchange.com/questions/171592/generate-random-numbers-with-linear-distribution

DoubleBetaDecayGenerators.Legacy — Module

Legacy code for generating events in the DECAY0 format.

DoubleBetaDecayGenerators.Legacy.dk0gen — Method

dk0gen(N; options)

Generate N events and dump them on file in the DECAY0 format.

DoubleBetaDecayGenerators.Legacy.momentum — Method

Convert energy to momentum (relativistic)

DoubleBetaDecayGenerators.Legacy.rand_2nbb — Method

rand_2nbb(data::TwoNuDBDData)

Generate the momenta of the two electrons emitted in two-neutrino double-beta decay, according to numerical calculations wrapped by 'data'. Units of keV.